services/captcha-ocr/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.4.0
pytesseract==0.3.13
cairosvg==2.7.1
ddddocr==1.5.6
